Switch to 32-bit hash entry keys
This is for added precision in detecting hash collisions.

Bench 5057873

+++ src/tt.h
@@ -97,7 +97,7 @@ public:
 
 private:
   friend struct TTEntry;
-  using EntryKey = uint16_t;
+  using EntryKey = uint32_t;
 
   size_t clusterCount;
   Cluster* table;
